Add periodic caching of filter config and states to disk
This feature stores gob encoded configuration and filter state
for each filter and stores it to disk before clearing the filter.
If cache exists and the configuration is the same when starting,
the filter states will be loaded back. This prevents state loss
if the service crashes or restarts, thereby preventing bursting
metrics after restart.